unsing ok

+++ Libft/ft_putnbr.c
@@ -6,15 +6,15 @@
 /*   By: ariard <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 16:24:18 by ariard            #+#    #+#             */
-/*   Updated: 2016/12/04 16:38:24 by ariard           ###   ########.fr       */
+/*   Updated: 2016/12/04 18:42:17 by ariard           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(long long int n)
 {
-	unsigned int	nb;
+	unsigned long long int	nb;
 
 	nb = n;
 	if (n < 0)
